[Refactor] RedisConfig, RedisUtils 파일 위치 변경
